6cdd77e2 COMP: Fixing warning

+++ Modules/Filtering/ImageGrid/test/itkInterpolateImagePointsFilterTest.cxx
@@ -164,7 +164,7 @@ int test2DInterpolateImagePointsFilter()
     {
     double value = outIter.Get();
     std::cout.width(10);
-    std::cout.precision( itk::Math::abs( std::log10( epsilon ) ) );
+    std::cout.precision( itk::Math::abs( unsigned( std::log10( epsilon ) ) ) );
     std::cout << "Checking image value: " << value << std::endl;
     if( !itk::Math::FloatAlmostEqual( value, truth[i], 10, epsilon ) )
       {
@@ -264,7 +264,7 @@ int test3DInterpolateImagePointsFilter()
   // Write home and let mom & dad know how we're doing.
   std::cout << "rmse of image is " << rmse << "\n ";
   double epsilon = 1e-7;
-  std::cout.precision( itk::Math::abs( std::log10( epsilon ) ) );
+  std::cout.precision( itk::Math::abs( unsigned( std::log10( epsilon ) ) ) );
   if ( !itk::Math::FloatAlmostEqual( rmse, (double)0 , 10, epsilon ) )
     {
     std::cout << "*** Error: rmse is larger than expected." << std::endl;
